pyproject.toml
[tool.poetry]
name = "table-tennis-game-annotater"
version = "0.0.1"
description = "Live annotation software for table tennis games"
authors = ["Kethan Vegunta <kethan@vegunta.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/kethan1/table-tennis-game-annotator"


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
opencv-python = "^4.6.0"
Pillow = "^9.2.0"
Flask = "^2.1.2"
torch = "^1.12.0"
pandas = "^1.4.3"
matplotlib = "^3.5.2"
PyYAML = "^6.0"
requests = "^2.28.1"
scipy = "^1.8.1"
torchvision = "^0.13.0"
protobuf = "<4.21.3"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
tqdm = "^4.64.0"

[tool.poe.tasks.watch]
shell = "npx parcel watch templates/*.html --no-source-maps --dist-dir dist"
interpreter = ["pwsh", "posix"]

[tool.poe.tasks.build]
shell = "npx parcel build templates/*.html --no-source-maps --dist-dir dist"
interpreter = ["pwsh", "posix"]

[tool.poe.tasks.server]
shell = "poetry run python app.py"
interpreter = ["pwsh", "posix"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
